// src/app/page/master/map-jabatan-pangkat/map-jabatan-pangkat.interface.ts
import { EventEmitter } from '@angular/core';

export interface MapJabatanPangkat {
	kdProduk?: string;
	kdEvent?: number;
	label?: any;
	value?: any;
	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
}